requirements.txt
# Core dependencies
requests>=2.28.0,<3.0.0
pytest>=7.0.0,<9.0.0
python-dotenv>=1.0.0,<2.0.0

# Reporting
allure-pytest>=2.13.0,<3.0.0

allure-python-commons